requirements.txt
numpy==2.2.0
tensorflow_intel==2.18.0
torch==2.5.1
transformers==4.46.3
